Refactored validation to be a class instantiated by the schema class.  Began development of new menu interface
